#28 - kafka
설정을 변경하고 예전설정에서 현재설정으로 doc를 참고하여 코드를 구현

+++ Moon/src/main/java/com/bigtyno/moon/model/event/AlarmEvent.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class AlarmEvent {
 
-    private Long receivedUserId;
+    private Integer receivedUserId;
     private AlarmType alarmType;
     private AlarmArgs alarmArgs;
